+++ CHANGELOG.md
@@ -31,6 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Ansible: Fix package download URL for DEB and RPM packages
 - Sendry Web: Fix sandbox POST handler not registered
+- Sendry Web: Fix DKIM deploy not updating domain configuration with KeyFile path
 
 ## [0.4.5] - 2026-01-29